Split out a `codec` module from `api-common`.
