+++ local-modules/api-client/ApiClient.js
@@ -2,7 +2,8 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { BaseKey, Codec, Message } from 'api-common';
+import { BaseKey, Message } from 'api-common';
+import { Codec } from 'codec';
 import { Logger } from 'see-all';
 import { TString } from 'typecheck';
 import { WebsocketCodes } from 'util-common';
