+++ local-modules/api-server/package.json
@@ -7,6 +7,7 @@
       "content-type": "^1.0.2",
 
       "api-common": "local",
+      "codec": "local",
       "hooks-server": "local",
       "see-all": "local",
       "server-env": "local",
