+++ local-modules/content-store-local/package.json
@@ -4,7 +4,7 @@
   "main": "main.js",
 
   "dependencies": {
-    "api-common": "local",
+    "codec": "local",
     "content-store": "local",
     "see-all": "local",
     "util-common": "local",
