+++ local-modules/doc-server/FileComplex.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { Codec } from 'api-common';
+import { Codec } from 'codec';
 import { BaseFile } from 'content-store';
 import { FrozenDelta } from 'doc-common';
 import { DEFAULT_DOCUMENT } from 'hooks-server';
